Split registration components into a new file
